Add Khronos Neutral tonemapper and add remapping and optional LogC LUT.
Testing showed that colours were skewed towards yellow. After some debugging,
it seems like this is caused by the CRT not actually being in 0-1 range,
and the Z position of the LUT did not reach 1. The remapping fixes this,
or at least gets much closer.

The LogC LUT was added for debugging, but it might be useful later
once colour adjustments are added.

Khronos Neutral Tonemapper is a colour preserving tonemapper that
works best with colours in 0-1 range. See more info at
https://github.com/KhronosGroup/ToneMapping/